Categorize buttons in html and change javascript
Instead of selecting all buttons and then applying
conditions, categorized the type of buttons in html
file as numbers, operations, equal key, dot key.
Now when a particular type of button is pressed
conditions are checked, if conditions met only then
certain actions occur. For example if a number is
pressed and no operation has occurred yet, the number
will be pushed in array1, if a number is pressed
and operation has been pressed then the number will
be pushed to array2, and so on.